requirements.txt
django>=4.2.0
plaid-python>=12.0.0
python-dotenv>=1.0.0 